change the type of the input parameter

+++ src/main/java/com/example/yp_skiresort/service/SkiResortSvc.java
@@ -7,6 +7,6 @@ import java.util.List;
 public interface SkiResortSvc {
     public abstract List<SkiResort> getSkiResortListByCountry(String country);
     public abstract List<SkiResort> getResortListByPriceRange(int max, int min);
-    public abstract List<SkiResort> getResortListBySlopeRating(int rating);
+    public abstract List<SkiResort> getResortListBySlopeRating(float rating);
     public abstract List<SkiResort> getResortListByPartialName(String resortName);
 }
